+++ app/views/api/v0/themes/extract.jbuilder
